% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{aest}
\alias{aest}
\title{计算样本的Aest统计量}
\usage{
aest(x, func)
}
\arguments{
\item{x}{一个数值向量，包含样本数据。}

\item{func}{一个函数，用于计算每个子集的统计量。这个函数应该接受一个数值向量作为输入，并返回一个数值。}
}
\value{
一个数值，表示Aest统计量。
}
\description{
这个函数用于计算给定样本的Aest统计量，这是一种用于估计样本数据中潜在异常值的统计量。
Aest统计量可以用于识别和处理数据集中的异常值。
}
\examples{
x <- c(7, 11, 15, 16, 20, 22, 24, 25, 29, 33)
func <- function(subset) mean(subset)
aest_result <- aest(x, func)
print(aest_result)
}
