% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{align_data}
\alias{align_data}
\title{Aligning Data in Completely Randomized Design
This function aligns the data for a completely randomized design by
adjusting each data point based on its group median and the overall median.}
\usage{
align_data(data, group_col, value_col)
}
\arguments{
\item{data}{A data frame containing the data}

\item{group_col}{The name of the column representing the group variable}

\item{value_col}{The name of the column representing the value variable}
}
\value{
A data frame with an additional column aligned_value containing the aligned data.
}
\description{
Aligning Data in Completely Randomized Design
This function aligns the data for a completely randomized design by
adjusting each data point based on its group median and the overall median.
}
\examples{
\donttest{
data <- data.frame(
  group = rep(1:3, each = 5),
  value = c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15)
)
align_data(data, "group", "value")
}
}
