% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{AlignedRankSP}
\alias{AlignedRankSP}
\title{使用Aligned-Rank Test检验裂区试验中的交互效应}
\usage{
AlignedRankSP(data, wholeplot, withinblock, random, response)
}
\arguments{
\item{data}{数据类型为dataframe的数据框}

\item{wholeplot}{wholeplot因子}

\item{withinblock}{withinblock因子}

\item{random}{作为随机效应的因子}

\item{response}{响应变量}
}
\value{
方差分析表格
}
\description{
通过剥离wholeplot因子的效应和withinblock因子的效应并转化为相应的秩，再使用anova
可以探究wholeplot因子和withinblock因子之间的交互效应
}
\examples{
\donttest{
AlignedRankSP(data = dog,
              wholeplot = "diabetes",
              withinblock = "method",
              random = "dogs",
              response = "turnover")
}
}
