% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{bca.interval}
\alias{bca.interval}
\title{计算BCA置信区间}
\usage{
bca.interval(a, boot, alpha, theta)
}
\arguments{
\item{a}{参数a}

\item{boot}{一个向量}

\item{alpha}{一个概率向量，通常包含一个或多个置信水平。}

\item{theta}{参数theta，用于计算booted样本中不大于theta的比例}
}
\value{
一个矩阵，包含BCA置信区间的下界和上界。
}
\description{
这个函数用于计算基于bootstrap方法的BCA（Bootstrap Percentile Confidence Interval）置信区间。
BCA置信区间是一种基于bootstrap方法的置信区间，它考虑了数据的整体分布。
}
\examples{
a <- 0.5
boot <- c(7, 11, 15, 16, 20, 22, 24, 25, 29, 33)
alpha <- 0.05
theta <- 20
bca_interval_result <- bca.interval(a, boot, alpha, theta)
print(bca_interval_result)
}
