% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{bootest_bias_var}
\alias{bootest_bias_var}
\title{Bootstrap方法估计偏差与方差}
\usage{
bootest_bias_var(x, fun, rep)
}
\arguments{
\item{x}{一个数值向量，包含原始数据。}

\item{fun}{一个函数，其偏差和方差将被估计。该函数应接受一个数值向量作为输入，并返回一个数值。}

\item{rep}{一个整数，指定Bootstrap过程的重复次数。}
}
\value{
一个数值向量，包含估计的偏差和方差。
}
\description{
这个函数使用Bootstrap方法来估计一个统计函数应用于给定数据集时的偏差和方差。
}
\examples{
x <- c(7, 11, 15, 16, 20, 22, 24, 25, 29, 33)
bootest_result <- bootest_bias_var(x, mean, 1000)
print(bootest_result)
}
