% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{bootest_mse}
\alias{bootest_mse}
\title{Bootstrap法估计均方误差（MSE）}
\usage{
bootest_mse(x, fun, rep)
}
\arguments{
\item{x}{一个数值向量，包含原始数据。}

\item{fun}{一个函数，其MSE将被估计。该函数应接受一个数值向量作为输入，并返回一个数值。}

\item{rep}{一个整数，指定自助法过程的重复次数。}
}
\value{
一个数值，代表函数\code{fun}应用于数据\code{x}时的估计均方误差（MSE）。
}
\description{
这个函数使用Bootstrap法来估计一个统计函数应用于给定数据集时的均方误差（MSE）。
}
\examples{
# 使用均值函数的例子
x <- rnorm(100)
mean_mse <- bootest(x, mean, 1000)
mean_mse

}
