% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{bootstrap_cor_ci}
\alias{bootstrap_cor_ci}
\title{相关系数Bootstrap区间估计}
\usage{
bootstrap_cor_ci(x, y, R = 1000, method = "bca")
}
\arguments{
\item{x}{向量，数据集的第一个变量}

\item{y}{向量，数据集的第二个变量}

\item{R}{整数，Bootstrap抽样次数}

\item{method}{字符串，区间估计的方法，可选"bca"或"perc"}
}
\value{
一个列表，包含原始相关系数和Bootstrap置信区间
}
\description{
相关系数Bootstrap区间估计
}
\examples{
\donttest{
x <- rnorm(100)
y <- 0.5 * x + rnorm(100)
bootstrap_cor_ci(x, y, R = 1000, method = "bca")
}
}
