% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_F_test.R
\name{bootstrap_F_test}
\alias{bootstrap_F_test}
\title{Bootstrap Sampling from Several Populations}
\usage{
bootstrap_F_test(data_list, n_bootstrap = 5000)
}
\arguments{
\item{data_list}{a list of data, each entry is data of one population}

\item{n_bootstrap}{the number of bootstrap samples that you want to construct}
}
\value{
F statistics based on the observed data, p value and F statistic based on the bootstrap samples
}
\description{
This is a function that can help you do bootstrap test of several populations using F statistics under same error variance.
}
\examples{
bootstrap_F_test(list(c1 = c(37.8, 37.1), c2 = c(37, 38.7,39.1), c3 = c(35.6, 32.4)),5000)
}
