% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{chi}
\alias{chi}
\title{计算卡方统计量}
\usage{
chi(x, s2)
}
\arguments{
\item{x}{一个数值向量，包含样本数据。}

\item{s2}{一个数值，代表假设方差。}
}
\value{
一个数值，表示卡方统计量。
}
\description{
这个函数用于计算给定样本的卡方统计量。
卡方统计量是一种用于比较样本方差与假设方差（s2）的统计量。
}
\examples{
x <- c(7, 11, 15, 16, 20, 22, 24, 25, 29, 33)
s2 <- 100
chi_result <- chi(x, s2)
print(chi_result)
}
