% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_boot.R
\name{confint.lb_boot}
\alias{confint.lb_boot}
\title{Calculate Confidence Intervals for lb_boot Objects}
\usage{
\method{confint}{lb_boot}(object, parm, level = 0.95, type = "percentile", ...)
}
\arguments{
\item{object}{An object of class \code{lb_boot}.}

\item{parm}{A specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.}

\item{level}{The confidence level required. Default is 0.95.}

\item{type}{Character string specifying the type of interval. Either "percentile" (default) or "normal".}

\item{...}{Additional arguments (not used).}
}
\value{
A matrix containing the confidence intervals.
}
\description{
Computes confidence intervals for bootstrap estimates using either the Percentile method or the Normal approximation.
}
