% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{cv}
\alias{cv}
\title{计算变异系数}
\usage{
cv(x)
}
\arguments{
\item{x}{一个数值向量，其变异系数将被计算。}
}
\value{
变异系数的值，以百分比形式表示。
}
\description{
这个函数计算输入向量的变异系数，它是标准差与均值的比例，乘以100转换为百分比形式。
}
\examples{
x <- c(7, 11, 15, 16, 20, 22, 24, 25, 29, 33)
cv_result <- cv(x)
print(cv_result)
}
