% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{FactDesignBoot}
\alias{FactDesignBoot}
\title{多因子试验设计中的bootstrap方法}
\usage{
FactDesignBoot(data, response, factor, formula, rep)
}
\arguments{
\item{data}{数据类型为dataframe的数据框}

\item{response}{目标变量，如"lifetime"}

\item{factor}{感兴趣的因子，如交互效应"temp:material"}

\item{formula}{用于anova的一个式子}

\item{rep}{抽样次数}
}
\value{
F统计量和bootstrapped p-value
}
\description{
多因子试验设计中的bootstrap方法
}
\examples{
\donttest{
FactDesignBoot(data = battery,
response = "lifetime",
factor = "temp:material",
formula = lifetime ~ temp * material,
rep = 1000)
}
}
