% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{fixed_x_bootstrap_cor_ci}
\alias{fixed_x_bootstrap_cor_ci}
\title{固定X的Bootstrap抽样}
\usage{
fixed_x_bootstrap_cor_ci(
  x,
  y,
  R = 1000,
  method = "bca",
  return_samples = FALSE
)
}
\arguments{
\item{x}{向量，数据集的第一个变量（自变量）}

\item{y}{向量，数据集的第二个变量（因变量）}

\item{R}{整数，Bootstrap抽样次数}

\item{method}{字符串，区间估计的方法，可选"bca"或"perc"}

\item{return_samples}{逻辑值，是否返回Bootstrap样本，默认为FALSE}
}
\value{
一个列表，包含原始相关系数、Bootstrap置信区间和（可选的）Bootstrap样本
}
\description{
固定X的Bootstrap抽样
}
\examples{
\donttest{
x <- rnorm(100)
y <- 0.5 * x + rnorm(100)
result <- fixed_x_bootstrap_cor_ci(x, y, R = 1000, method = "bca", return_samples = FALSE)
str(result)
}
}
