% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lattice_ordered_test.R
\name{lattice_ordered_test}
\alias{lattice_ordered_test}
\title{Function to perform the lattice-ordered alternatives test}
\usage{
lattice_ordered_test(data, A_levels, B_levels, n_permutations = 1000)
}
\arguments{
\item{data}{the data you want to perform the test}

\item{A_levels}{A level}

\item{B_levels}{B level}

\item{n_permutations}{permutation time}
}
\value{
observed_TL, p value
}
\description{
Function to perform the lattice-ordered alternatives test
}
\examples{
lattice_ordered_test(matrix(c(1,2,3,4,5,6,7,8,9),nrow = 3,byrow = TRUE,dimnames=list(c("P1", "P2", "P3"), c("Week 1", "Week 2", "Week 4"))), c("P1", "P2", "P3"), c("Week 1", "Week 2", "Week 4"))
}
