% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_art.R
\name{lb_art}
\alias{lb_art}
\title{Aligned Rank Transform (ART) for Split-Plot Designs}
\usage{
lb_art(data, response, whole_plot, sub_plot, block)
}
\arguments{
\item{data}{A data frame containing the experimental data.}

\item{response}{Character. The name of the response variable column.}

\item{whole_plot}{Character. The name of the whole-plot factor.}

\item{sub_plot}{Character. The name of the sub-plot (within-block) factor.}

\item{block}{Character. The name of the block/subject identifier (e.g., "farm" or "plotID").}
}
\value{
An object of class \code{"anova"} containing the F-test results for the interaction.
}
\description{
Performs the Aligned Rank Transform test for interaction effects in split-plot designs.
This method aligns the data by removing main effects and then applying a rank-based ANOVA
on the residuals, as described in Higgins (Section 9.2).
}
