% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lb_boot.R
\name{print.lb_boot}
\alias{print.lb_boot}
\title{Print method for lb_boot objects}
\usage{
\method{print}{lb_boot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lb_boot}.}

\item{...}{Additional arguments (not used).}
}
\description{
Nicely formats the output of the lb_boot function, showing bias and standard error.
}
