% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{resboot}
\alias{resboot}
\title{多元线性回归中的residuals bootstrap}
\usage{
resboot(data, formula, rep, conf_level)
}
\arguments{
\item{data}{数据类型为dataframe的数据框}

\item{formula}{类似于y ~ x1 + x2的表达式}

\item{rep}{重抽样次数}

\item{conf_level}{置信水平}
}
\value{
返回系数估计值和对应的置信区间
}
\description{
多元线性回归中的residuals bootstrap是一种半参数方法，所依赖的假设为误差独立且同方差。
}
\examples{
\donttest{
resboot(data = bodyfat,
        formula = bodyfat ~ Weight + Chest + Abdomen,
        rep = 1000,
        conf_level = 0.95)
}

}
