% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{stt}
\alias{stt}
\title{计算样本统计量与已知参数之间的差异的标准化值}
\usage{
stt(x, theta)
}
\arguments{
\item{x}{一个数值向量，包含样本数据。}

\item{theta}{一个数值，代表已知的参数值。}
}
\value{
一个数值，表示样本统计量与已知参数之间的差异的标准化值。
}
\description{
这个函数用于计算一个给定样本的统计量与一个已知参数之间的差异的标准化值。
这个标准化值可以帮助我们理解样本统计量与已知参数值之间的差异，并将其标准化到与样本大小无关的量度。
在假设检验中，这个标准化值可以用于计算统计检验的统计量，如t统计量或z统计量。
}
\examples{
x <- c(7, 11, 15, 16, 20, 22, 24, 25, 29, 33)
theta <- 18
stt_result <- stt(x, theta)
print(stt_result)
}
